14-content-moderation/requirements.txt
transformers
Pillow
torch --extra-index-url https://download.pytorch.org/whl/cu119